Refactoring more repair on the client side
