+++ client/src/repair/ActionInfo.ts
@@ -1,0 +1,6 @@
+import { Range } from 'vscode';
+
+export default interface ActionInfo {
+	replacementText: string,
+	range: Range
+}
