--- client/src/repair/SingleCopyRepair.ts
@@ -1,121 +1,0 @@
-import {
-  CancellationToken,
-  CodeAction,
-  CodeActionContext,
-  CodeActionProvider,
-  Range,
-  Selection,
-  TextDocument,
-} from "vscode";
-import {
-  createAction,
-  getNewline,
-  getNumberOfCharsForNewline,
-  isNodeProject,
-} from "./utils";
-
-const SINGLE_COPY_MSG = "Add a second COPY instruction.";
-const SINGLE_COPY_CODE = "R:SINGLECOPY";
-
-export default class SingleCopyRepair
-  implements CodeActionProvider<CodeAction>
-{
-  provideCodeActions(
-    document: TextDocument,
-    range: Range | Selection,
-    context: CodeActionContext,
-    token: CancellationToken
-  ): CodeAction[] {
-    const actions: CodeAction[] = [];
-
-    for (const diagnostic of context.diagnostics) {
-      if (diagnostic.code !== SINGLE_COPY_CODE) continue;
-
-      const documentText = document.getText();
-
-      const originalCopyIndex = documentText.indexOf("COPY");
-
-      const newlineChar = getNewline();
-
-      const firstNewlineAfterCopy = documentText.indexOf(
-        newlineChar,
-        originalCopyIndex
-      );
-
-      const originalCopyContent = documentText
-        .substring(originalCopyIndex, firstNewlineAfterCopy)
-        .replace("  ", " ");
-
-      const copyComponents = originalCopyContent.split(" ");
-
-      if (!copyComponents || copyComponents.length < 3) continue;
-
-      const secondArg = copyComponents[2];
-
-      const newLineCharLength = getNumberOfCharsForNewline();
-      const indexAfterOriginalCopy =
-        documentText.indexOf(newlineChar, originalCopyIndex) +
-        newLineCharLength;
-
-      const cmdIndex = documentText.indexOf("CMD");
-      const entrypointIndex = documentText.indexOf("ENTRYPOINT");
-      const lastInstructionIndex = cmdIndex !== -1 ? cmdIndex : entrypointIndex;
-      const indexBeforeLastInstruction =
-        lastInstructionIndex - newLineCharLength;
-
-      const rangeToBeMaintained = new Range(
-        document.positionAt(indexAfterOriginalCopy),
-        document.positionAt(indexBeforeLastInstruction)
-      );
-      const textToBeMaintained = document.getText(rangeToBeMaintained);
-
-      const isNode = isNodeProject(document);
-
-      const chownText = getChownText(documentText);
-
-      const firstCopy =
-        "COPY " +
-        (isNode ? "package*.json" : "requirements.txt") +
-        " " +
-        secondArg +
-        (secondArg.endsWith("/") ? "" : "/") +
-        newlineChar;
-      const secondCopy = `${newlineChar}COPY ${chownText}. .${newlineChar}`;
-      const replacementText = firstCopy + textToBeMaintained + secondCopy;
-
-      const rangeToBeReplaced = new Range(
-        document.positionAt(originalCopyIndex),
-        document.positionAt(indexBeforeLastInstruction)
-      );
-
-      const action = createAction(
-        SINGLE_COPY_MSG,
-        replacementText,
-        document,
-        rangeToBeReplaced
-      );
-
-      actions.push(action);
-    }
-
-    return actions;
-  }
-}
-
-function getChownText(documentText: string): string {
-  const userIndex = documentText.indexOf("USER");
-
-  if (userIndex === -1) return "";
-
-  const newlineChar = getNewline();
-
-  const newlineAfterUserIndex = documentText.indexOf(newlineChar, userIndex);
-
-  const userStartIndex = userIndex + "USER ".length;
-
-  const user = documentText.substring(userStartIndex, newlineAfterUserIndex);
-
-  const chownText = `--chown=${user}:${user} `;
-
-  return chownText;
-}
